Orders_API/requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.20
mysqlclient==2.2.0
